+++ src/main/kotlin/kr/io/team/loop/task/domain/model/TaskCommand.kt
@@ -3,7 +3,6 @@ package kr.io.team.loop.task.domain.model
 import kotlinx.datetime.LocalDate
 import kr.io.team.loop.common.domain.GoalId
 import kr.io.team.loop.common.domain.MemberId
-import kr.io.team.loop.common.domain.TaskId
 
 sealed interface TaskCommand {
     data class Create(
