+++ src/main/kotlin/kr/io/team/loop/task/domain/repository/TaskRepository.kt
@@ -1,9 +1,9 @@
 package kr.io.team.loop.task.domain.repository
 
 import kr.io.team.loop.common.domain.GoalId
-import kr.io.team.loop.common.domain.TaskId
 import kr.io.team.loop.task.domain.model.Task
 import kr.io.team.loop.task.domain.model.TaskCommand
+import kr.io.team.loop.task.domain.model.TaskId
 import kr.io.team.loop.task.domain.model.TaskQuery
 
 interface TaskRepository {
