+++ src/main/kotlin/kr/io/team/loop/task/infrastructure/persistence/ExposedTaskRepository.kt
@@ -2,9 +2,9 @@ package kr.io.team.loop.task.infrastructure.persistence
 
 import kr.io.team.loop.common.domain.GoalId
 import kr.io.team.loop.common.domain.MemberId
-import kr.io.team.loop.common.domain.TaskId
 import kr.io.team.loop.task.domain.model.Task
 import kr.io.team.loop.task.domain.model.TaskCommand
+import kr.io.team.loop.task.domain.model.TaskId
 import kr.io.team.loop.task.domain.model.TaskQuery
 import kr.io.team.loop.task.domain.model.TaskStatus
 import kr.io.team.loop.task.domain.model.TaskTitle
