+++ src/test/kotlin/kr/io/team/loop/task/domain/model/TaskTest.kt
@@ -5,7 +5,7 @@ import io.kotest.matchers.shouldBe
 import kotlinx.datetime.LocalDate
 import kr.io.team.loop.common.domain.GoalId
 import kr.io.team.loop.common.domain.MemberId
-import kr.io.team.loop.common.domain.TaskId
+import kr.io.team.loop.task.domain.model.TaskId
 import java.time.Instant
 
 class TaskTest :
